fix: a minor hack to get control over the caches
Sometimes items that should have been removed from the caches end up back in them.

This checks the rendereditem and the graphicsContainer widht. If they are not
the same as Garden ItemWidht, the cached version is discarded and recreated.

The renderedItem width makes sure the color gets sized properly.
The graphicsContainer makes sure the text get placed properly.